requirements_optimized.txt
# Dependencias principales
fastapi[standard]
tensorflow==2.19.0
numpy==2.1.3
pillow==11.2.1
uvicorn==0.34.2

# Dependencias para explicabilidad (si planeas usarlas)
# shap==0.48.0
# matplotlib==3.10.6
